bot/requirements.txt
PyYAML      
numpy
pandas
scikit-learn
torch
transformers
pytz
requests
aiohttp
matplotlib
seaborn
ta
optuna
tensorboard
textblob
metaapi-cloud-sdk
tensorflow
yfinance
flask-socketio
forex-python
